pyproject.toml
[tool.poetry]
name = "mutt"
version = "0.1.0"
description = ""
authors = ["Matias <modandyxx@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
psycopg2 = "^2.9.7"
regex = "^2023.8.8"
argparse = "^1.4.0"
requests = "^2.31.0"
python-dateutil = "^2.8.2"
logging = "^0.4.9.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
